Moved Group related files to Group/ folder in filesystem
